Revert "fix: refactor folders and fix app not running in android"
