--- android/app/src/main/res/values-night/colors.xml
@@ -1,1 +1,0 @@
-<resources/>
