--- android/app/src/main/res/values/strings.xml
@@ -1,6 +1,0 @@
-<resources>
-  <string name="app_name">CapturGO</string>
-  <string name="expo_splash_screen_resize_mode" translatable="false">contain</string>
-  <string name="expo_splash_screen_status_bar_translucent" translatable="false">false</string>
-  <string name="expo_system_ui_user_interface_style" translatable="false">automatic</string>
-</resources>
