+++ ios/Podfile
@@ -13,8 +13,8 @@ install! 'cocoapods',
 
 prepare_react_native_project!
 
-# @generated begin @rnmapbox/maps-rnmapboxmapsimpl - expo prebuild (DO NOT MODIFY) sync-da39a3ee5e6b4b0d3255bfef95601890afd80709
-
+# @generated begin @rnmapbox/maps-rnmapboxmapsimpl - expo prebuild (DO NOT MODIFY) sync-0827c5766d67698b43a606eed3991302021cf996
+$RNMapboxMapsDownloadToken = 'sk.eyJ1IjoiYmh1bmFraXQiLCJhIjoiY21iMGtoNW40MHMxMDJqcjBscXE1aHN4MCJ9.Rt0FTUZVapci4ba_RzSvng'
 # @generated end @rnmapbox/maps-rnmapboxmapsimpl
 target 'CapturGO' do
   use_expo_modules!
