+++ package.json
@@ -12,39 +12,39 @@
   },
   "dependencies": {
     "@expo/vector-icons": "^14.1.0",
-    "@react-native-async-storage/async-storage": "2.1.2",
-    "@react-native-picker/picker": "2.11.1",
+    "@react-native-async-storage/async-storage": "^2.1.2",
+    "@react-native-picker/picker": "^2.11.0",
     "@react-navigation/bottom-tabs": "^7.3.10",
     "@react-navigation/elements": "^2.3.8",
     "@react-navigation/native": "^7.1.6",
     "@rnmapbox/maps": "^10.1.39",
     "@supabase/supabase-js": "^2.49.8",
     "@turf/turf": "^7.2.0",
-    "expo": "53.0.19",
+    "expo": "~53.0.9",
     "expo-blur": "~14.1.4",
     "expo-constants": "~17.1.6",
     "expo-device": "~7.1.4",
     "expo-font": "~13.3.1",
     "expo-haptics": "~14.1.4",
-    "expo-image": "~2.3.2",
-    "expo-linking": "~7.1.7",
-    "expo-location": "~18.1.6",
+    "expo-image": "~2.1.7",
+    "expo-linking": "~7.1.5",
+    "expo-location": "~18.1.5",
     "expo-maps": "~0.11.0",
-    "expo-router": "~5.1.3",
-    "expo-splash-screen": "~0.30.10",
+    "expo-router": "~5.0.6",
+    "expo-splash-screen": "~0.30.8",
     "expo-status-bar": "~2.2.3",
     "expo-symbols": "~0.4.4",
-    "expo-system-ui": "~5.0.10",
-    "expo-task-manager": "~13.1.6",
+    "expo-system-ui": "~5.0.7",
+    "expo-task-manager": "~13.1.5",
     "expo-tracking-transparency": "^5.2.4",
-    "expo-web-browser": "~14.2.0",
+    "expo-web-browser": "~14.1.6",
     "react": "19.0.0",
     "react-dom": "19.0.0",
-    "react-native": "0.79.5",
+    "react-native": "0.79.2",
     "react-native-gesture-handler": "~2.24.0",
     "react-native-reanimated": "~3.17.4",
     "react-native-safe-area-context": "5.4.0",
-    "react-native-screens": "~4.11.1",
+    "react-native-screens": "~4.10.0",
     "react-native-url-polyfill": "^2.0.0",
     "react-native-web": "~0.20.0",
     "react-native-webview": "13.13.5"
